handle `null` for Workspace::representation
It seems that on newly created, empty workspaces the representation can be `null`, with this change the representation gets deserialized to an empty string in that case.
Another option would be to switch to Option<String>, which I haven't done here to preserve backwards compatibility.

Additionally this changes from requiring `serde_derive` directly towards using the `derive` feature of `serde`, because the compiler didn't like `use`ing a trait and macro of the same name.