fix: don't crash when connecting and neither i3 nor sway is installed
I maintain a project that uses swayipc to communicate to the window
manager.

Recently, I added support for hyprland and I got [a report](https://github.com/pierrechevalier83/workstyle/issues/49) that my
executable was crashing.
The crash was originating from [a call to `Connection::new()`](https://github.com/pierrechevalier83/workstyle/blob/b85006e6af8e45fc36d613ac92b51699646a4de8/src/window_manager.rs#L263), and the
user confirmed that they didn't have Sway or I3 installed on their
machine.

I had assumed that this call would err if it couldn't connect to the
window manager (which it does as long as i3 or sway is installed),
since it returned a `Result`, so I used it to decide which window
manager to connect to: try sway or i3 first and if that fails, try
hyprland.

Now I've read the code here, I could fix it on my side by basically
re-implementing the logic inside `get_socketpath` in my crate, but I'd
rather swayipc behaved as I had originally thought it would if you're OK
with the change.