Allow empty field `status_command`
Do not require the field `status_command` to be a String type.
Make it possible for deserialize to handle `null` value.

If the field `status_command` is not set in sway, the returned value in
`null`. This can be seen with `swaymsg -t get_bar_config bar-0`.
In case of `null` the test `get_bar_ids_and_one_config()` fails.

After making the field optional, serde does not crash and the test
passes.